+++ changelog.md
@@ -6,6 +6,8 @@ with 1.0.0beta.
 ## \[1.1.6] - Unreleased
 
 - Add: Change email address. **Requires backend version 1.1.3** (no tag yet)
+- Improve: Default link color is now the brand blue.
+- Improve: Link underline hover transition.
 
 ## \[1.1.5] - 2024-04-22